Decompile the rest of code_08002BA8.s
